Finished refactoring home page, now is fully responive.
